rename to xx_decompose; also fix a test error
