+++ qiskit/quantum_info/synthesis/xx_decompose/paths.py
