+++ qiskit/quantum_info/synthesis/xx_decompose/polytopes.py
