+++ qiskit/quantum_info/synthesis/xx_decompose/utilities.py
