+++ test/python/quantum_info/xx_decompose/test_circuits.py
@@ -11,7 +11,7 @@
 # that they have been altered from the originals.
 
 """
-Tests for qiskit-terra/qiskit/quantum_info/synthesis/rzx_decompose/circuits.py .
+Tests for qiskit-terra/qiskit/quantum_info/synthesis/xx_decompose/circuits.py .
 """
 
 from operator import itemgetter
@@ -25,7 +25,7 @@ from qiskit.circuit import QuantumCircuit
 from qiskit.circuit.library import RZGate
 import qiskit.quantum_info.operators
 from qiskit.quantum_info.synthesis.weyl import weyl_coordinates
-from qiskit.quantum_info.synthesis.rzx_decompose.circuits import (
+from qiskit.quantum_info.synthesis.xx_decompose.circuits import (
     decompose_xxyy_into_xxyy_xx,
     xx_circuit_step,
 )
